secao-03_ice-breaker/pyproject.toml
[project]
name = "secao-03-ice-breaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-groq>=0.3.1",
    "langchainhub>=0.1.21",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
